requirements.txt
webrtcvad==2.0.10
sounddevice==0.4.7
numpy==2.1.2
scipy==1.13.1
pydub==0.25.1
python-dotenv==1.0.1

requests>=2.31.0
